requirements.txt
fastapi==0.115.12
numpy==1.23.5
onnxruntime==1.22.0
opencv_python==4.8.0.76
opencv_python_headless==4.11.0.86
pandas==1.5.3
PyYAML==6.0.2
segmentation_models_pytorch==0.5.0
torch==2.7.0
uvicorn==0.34.2
